refactor cog status to presence
